improve command lin e interface structure
