+++ src/CLI/BaseCommand.php
@@ -1,13 +1,11 @@
 <?php
 
-namespace Pionia\command;
+namespace Pionia\CLI;
 
 use Pionia\core\Pionia;
 use Pionia\database\Connection;
-use Pionia\Logging\PioniaLogger;
 use Porm\core\Database;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\HttpKernel\Log\Logger;
 
 /**
  * This is the base command class, it extends the Symfony console command class and provides some helper methods
