+++ src/CLI/commands/CreateMiddleware.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Pionia\command\commands;
+namespace Pionia\CLI\commands;
 
 use Pionia\codegens\AuthBackend;
-use Pionia\command\BaseCommand;
+use Pionia\codegens\Middleware;
+use Pionia\CLI\BaseCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -14,18 +15,18 @@ use Symfony\Component\Console\Output\OutputInterface;
  *
  * @author [Jet - ezrajet9@gmail.com](https://www.linkedin.com/in/jetezra/)
  */
-class CreateAuthenticationBackend extends BaseCommand
+class CreateMiddleware extends BaseCommand
 {
-    protected static string $title = 'Adds a new authentication backend';
-    protected static string $description = 'add authentication backend';
-    protected static string $name = 'addauth';
+    protected static string $title = 'Adds a new middleware to pionia app';
+    protected static string $description = 'add middleware backend';
+    protected static string $name = 'addmidd';
     protected function configure(): void
     {
         $this
             ->setName($this::$name)
-            ->setDescription('Creates a '.$this::base()::$name.' authentication backend')
-            ->addArgument('name', InputArgument::REQUIRED, 'The name of the authentication backend')
-            ->setHelp('Create a new authentication backend of a '.$this::base()::$name.' application in app/authenticationBackends');
+            ->setDescription('Creates a '.$this::base()::$name.' middleware')
+            ->addArgument('name', InputArgument::REQUIRED, 'The name of the meddleware to create')
+            ->setHelp('Create a new middleware of a '.$this::base()::$name.' application in app/middlewares');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
@@ -34,7 +35,7 @@ class CreateAuthenticationBackend extends BaseCommand
 
         $output->writeln("Generating $service_name...");
 
-        $service = new AuthBackend($service_name, $output);
+        $service = new Middleware($service_name, $output);
         $service->generate();
 
         return Command::SUCCESS;
