+++ src/CLI/commands/StartServer.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Pionia\command\commands;
+namespace Pionia\CLI\commands;
 
-use Pionia\command\BaseCommand;
+use Pionia\CLI\BaseCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
